requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pyspellchecker==0.8.2
pytesseract==0.3.13
Werkzeug==3.1.3
